Logic for pragma version replacement in python function, fix undeterministic behavior

+++ script/util/process_briefcase_files.py
@@ -4,7 +4,7 @@ import sys
 
 
 def parse_flattened_file(file_path):
-    with open(file_path, 'r') as file:
+    with open(file_path, "r") as file:
         lines = file.readlines()
 
     main_pragma_version = None
@@ -17,11 +17,14 @@ def parse_flattened_file(file_path):
                 break
 
     code_blocks = []
-    current_block = {'source': None, 'content': '',
-                     'pragma': main_pragma_version}
+    current_block = {"source": None, "content": "", "pragma": main_pragma_version}
 
     for line in lines:
-        if line.startswith("import {") or line.startswith("import \"") or line.startswith("import '"):
+        if (
+            line.startswith("import {")
+            or line.startswith('import "')
+            or line.startswith("import '")
+        ):
             continue
 
         # License is taken from source file, so can be ignored
@@ -41,50 +44,53 @@ def parse_flattened_file(file_path):
             if current_block["source"]:
                 code_blocks.append(current_block)
                 current_block = {
-                    'source': line.strip(), 'content': '', 'pragma': main_pragma_version, 'imports': []}
+                    "source": line.strip(),
+                    "content": "",
+                    "pragma": main_pragma_version,
+                    "imports": [],
+                }
             else:
-                current_block['source'] = line.strip()
+                current_block["source"] = line.strip()
 
             continue
 
-        current_block['content'] += line
+        current_block["content"] += line
 
     code_blocks.append(current_block)
 
     return code_blocks
 
 
-def get_pragma_and_license_from_source_file(source_file, parse_pragma_versions=False):
-    if parse_pragma_versions:
-        pragma_pattern = re.compile(r"^\s*pragma\s+[^;]+;")
-    else:
-        pragma_pattern = re.compile(r"^\s*pragma\s+(?!solidity)[^;]+;")
+def get_pragma_and_license_from_source_file(source_file):
+    pragma_version_pattern = re.compile(r"^\s*pragma\s+solidity\s+[^;]+;")
+    pragma_additional_pattern = re.compile(r"^\s*pragma\s+(?!solidity\s)[^;]+;")
+    license_pattern = re.compile(r"^\s*//\s*SPDX-License-Identifier:\s*([^\s]+)")
 
-    license_pattern = re.compile(
-        r"^\s*//\s*SPDX-License-Identifier:\s*([^\s]+)")
-
-    additional_pragma_lines = []
+    pragma_version_line = None
+    pragma_additional_lines = []
     license = None
 
-    with open(source_file, 'r') as file:
+    with open(source_file, "r") as file:
         for line in file:
             if license_pattern.match(line):
                 license = line.strip()
-            if pragma_pattern.match(line):
-                additional_pragma_lines.append(line.strip())
+            if pragma_version_pattern.match(line):
+                pragma_version_line = line.strip()
+            if pragma_additional_pattern.match(line):
+                pragma_additional_lines.append(line.strip())
 
-    return additional_pragma_lines, license
+    return pragma_version_line, pragma_additional_lines, license
 
 
 def remove_comments_and_assembly(code):
     # Remove single-line comments
-    code = re.sub(r'//.*', '', code)
+    code = re.sub(r"//.*", "", code)
     # Remove multi-line comments
-    code = re.sub(r'/\*.*?\*/', '', code, flags=re.DOTALL)
+    code = re.sub(r"/\*.*?\*/", "", code, flags=re.DOTALL)
 
     # Pattern to find assembly blocks (with optional annotation like ("memory-safe"))
     # Since Python allows no recursive regex, we need to be greedy for the last '}' it can find and find the correct end of block ourselves
-    assembly_pattern = r'assembly\s*\(?.*?\)?\s*\{(.*\n?)*\}'
+    assembly_pattern = r"assembly\s*\(?.*?\)?\s*\{(.*\n?)*\}"
     match = re.search(assembly_pattern, code, re.DOTALL)
 
     while match:
@@ -96,9 +102,9 @@ def remove_comments_and_assembly(code):
 
         # Traverse the matched assembly block to handle nested braces
         for i in range(start_idx, end_idx):
-            if code[i] == '{':
+            if code[i] == "{":
                 open_braces += 1
-            elif code[i] == '}':
+            elif code[i] == "}":
                 open_braces -= 1
                 if open_braces == 0:
                     # Found the matching closing brace
@@ -126,16 +132,15 @@ def get_imported_identifiers_from_code(code):
     matches_with_braces = re.findall(import_with_braces_regex, code)
 
     for match in matches_with_braces:
-        identifiers = match[0].split(',')
+        identifiers = match[0].split(",")
         identifiers = [identifier.strip() for identifier in identifiers]
         imported_identifiers.extend(identifiers)
 
-    matches_without_braces = re.findall(
-        import_without_braces_regex, code)
+    matches_without_braces = re.findall(import_without_braces_regex, code)
 
     for match in matches_without_braces:
         filename = os.path.basename(match)
-        class_name = filename.replace('.sol', '')
+        class_name = filename.replace(".sol", "")
         class_name = class_name[0].upper() + class_name[1:]
         imported_identifiers.append(class_name)
 
@@ -143,7 +148,7 @@ def get_imported_identifiers_from_code(code):
 
 
 def get_imported_identifiers_in_source_file(source_file):
-    with open(source_file, 'r') as file:
+    with open(source_file, "r") as file:
         sol_content = file.read()
         imported_identifiers = get_imported_identifiers_from_code(sol_content)
 
@@ -154,8 +159,8 @@ def parse_content_for_identifier_definitions(content):
     identifiers = []
     functions = []
     # Define regex patterns for contracts, structs, interfaces, and libraries
-    identifier_pattern = r'\b(?:interface|contract|type|library|struct)\s+(\w+)\b'
-    function_pattern = r'\b(?:function)\s+(\w+)\b'
+    identifier_pattern = r"\b(?:interface|contract|type|library|struct)\s+(\w+)\b"
+    function_pattern = r"\b(?:function)\s+(\w+)\b"
     # Remove comments
     content = remove_comments_and_assembly(content)
 
@@ -165,20 +170,26 @@ def parse_content_for_identifier_definitions(content):
     for line in content.splitlines():
         # Only handle top-level definitions
         if scope_level == 0 and line:
-            identifier_match = re.search(
-                identifier_pattern, line, re.MULTILINE)
-            function_match = re.search(function_pattern, line, re.MULTILINE,)
+            identifier_match = re.search(identifier_pattern, line, re.MULTILINE)
+            function_match = re.search(
+                function_pattern,
+                line,
+                re.MULTILINE,
+            )
             if identifier_match:
                 identifiers.append(identifier_match.group(1))
                 identifier_match = re.search(
-                    identifier_pattern, line, re.MULTILINE,)
+                    identifier_pattern,
+                    line,
+                    re.MULTILINE,
+                )
             if function_match:
                 functions.append(function_match.group(1))
 
         for char in line:
-            if char == '{':
+            if char == "{":
                 scope_level += 1
-            elif char == '}':
+            elif char == "}":
                 scope_level -= 1
 
     return identifiers, functions
@@ -187,14 +198,15 @@ def parse_content_for_identifier_definitions(content):
 def get_used_identifiers(code, identifiers):
     code = remove_comments_and_assembly(code)
 
-    self_declared_identifiers, _ = parse_content_for_identifier_definitions(
-        code)
+    self_declared_identifiers, _ = parse_content_for_identifier_definitions(code)
     # It's possible an identifier was declared twice in flattened file, it shouldn't try to import a contract with its own name
     identifiers = [
-        identifier for identifier in identifiers if identifier not in self_declared_identifiers]
+        identifier
+        for identifier in identifiers
+        if identifier not in self_declared_identifiers
+    ]
 
-    identifiers_pattern = '|'.join(re.escape(identifier)
-                                   for identifier in identifiers)
+    identifiers_pattern = "|".join(re.escape(identifier) for identifier in identifiers)
     pattern = rf"(?<![\w.])({identifiers_pattern})(?!\w)"
     # Compile with verbose mode to make the pattern easier to read
     pattern_compiled = re.compile(pattern, re.VERBOSE | re.DOTALL)
@@ -211,14 +223,19 @@ def get_used_identifiers(code, identifiers):
 
 def get_relative_path(src_path, dest_path):
     relative_path = os.path.relpath(src_path, os.path.dirname(dest_path))
-    if not relative_path.startswith('.'):
+    if not relative_path.startswith("."):
         relative_path = os.path.join(".", relative_path)
     return relative_path
 
 
-def write_file(path, code_block, license, pragma, identifiers, functions, imports):
+def write_file(
+    path, code_block, license, pragma, identifiers, functions, imports, overwrite=False
+):
+    if not overwrite and os.path.exists(path):
+        return
+
     os.makedirs(os.path.dirname(path), exist_ok=True)
-    with open(path, 'w') as file:
+    with open(path, "w") as file:
         if license:
             file.write(license + "\n")
         file.write(pragma + "\n\n")
@@ -226,18 +243,17 @@ def write_file(path, code_block, license, pragma, identifiers, functions, import
         for dest_path in identifiers:
             if dest_path != path:
                 identifier_names = identifiers[dest_path]
-                used_identifiers = get_used_identifiers(
-                    code_block, identifier_names)
+                used_identifiers = get_used_identifiers(code_block, identifier_names)
 
-                used_functions = list(
-                    set(functions[dest_path]).intersection(imports))
+                used_functions = list(set(functions[dest_path]).intersection(imports))
                 if used_functions:
                     used_identifiers += used_functions
 
                 if used_identifiers:
                     relative_path = get_relative_path(dest_path, path)
                     import_statement = create_import_statement(
-                        used_identifiers, relative_path)
+                        used_identifiers, relative_path
+                    )
                     file.write(import_statement)
 
         file.write("\n" + code_block.strip())
@@ -247,7 +263,23 @@ def create_import_statement(identifier_names, relative_path):
     return f'import {{{", ".join(identifier_names)}}} from "{relative_path}";\n'
 
 
-def process_file(flattened_file, flattened_dir, contracts_dir):
+def check_if_should_overwrite_pragma(flattened_file_path):
+    if re.match(r"(.*)/(interface|interfaces|types)/(.*)", flattened_file_path):
+        return True
+
+    flattened_file_name = os.path.basename(flattened_file_path)
+    if re.match(r"^I[A-Z]", flattened_file_name):
+        return True
+
+    return False
+
+# Normalize pragma version statement for potential whitespace so comparison is safer
+def normalize_pragma_version(pragma_version):
+    match = re.match(r"^\s*pragma\s+solidity\s+([^;]+);", pragma_version)
+    return f"pragma solidity {match.group(1)};" if match else pragma_version
+
+
+def process_file(flattened_file, contracts_dir, target_dir):
     code_blocks = parse_flattened_file(flattened_file)
     pragma = {}
     license = {}
@@ -259,16 +291,16 @@ def process_file(flattened_file, flattened_dir, contracts_dir):
     # get packages under src/pkgs
     pkgs_dir = os.path.join(contracts_dir, "src", "pkgs")
     available_packages = [
-        pkg for pkg in os.listdir(pkgs_dir)
+        pkg
+        for pkg in os.listdir(pkgs_dir)
         if os.path.isdir(os.path.join(pkgs_dir, pkg))
     ]
 
-    flattened_file_match = re.match(
-        r'(.*)/(interface|interfaces|types)/(.*)', flattened_file)
+    overwrite_pragma = check_if_should_overwrite_pragma(flattened_file)
 
     for block in code_blocks:
-        if block['source']:
-            source_file = block['source'].replace("// ", "")
+        if block["source"]:
+            source_file = block["source"].replace("// ", "")
         else:
             break
 
@@ -283,12 +315,13 @@ def process_file(flattened_file, flattened_dir, contracts_dir):
             source_file = os.path.join("src", "pkgs", *path_parts)
 
         interface_match = re.match(
-            r'src/pkgs/(.*)/(src|contracts)/(interface|interfaces)/(.*)', source_file)
+            r"src/pkgs/(.*)/(src|contracts)/(interface|interfaces)/(.*)", source_file
+        )
         lib_types_match = re.match(
-            r'src/pkgs/(.*)/(src|contracts)/(libraries|types)/(.*)', source_file)
-        external_lib_match = re.match(r'(.*?)lib/(.*)', source_file)
-        other_match = re.match(
-            r'src/pkgs/(.*)/(src|contracts)/(.*)', source_file)
+            r"src/pkgs/(.*)/(src|contracts)/(libraries|types)/(.*)", source_file
+        )
+        external_lib_match = re.match(r"(.*?)lib/(.*)", source_file)
+        other_match = re.match(r"src/pkgs/(.*)/(src|contracts)/(.*)", source_file)
 
         if interface_match:
             package_name = interface_match.group(1)
@@ -299,63 +332,77 @@ def process_file(flattened_file, flattened_dir, contracts_dir):
             subdir = lib_types_match.group(3)
             rest_of_path = lib_types_match.group(4)
         elif external_lib_match:
-            package_name = ''
+            package_name = ""
             subdir = "lib-external"
             rest_of_path = external_lib_match.group(2)
         elif other_match:
             package_name = other_match.group(1)
-            subdir = ''
+            subdir = ""
             rest_of_path = other_match.group(3)
         else:
             continue
 
-        content = block['content']
+        content = block["content"]
 
         dest_path = os.path.join(
-            flattened_dir, package_name, subdir, rest_of_path)
+            contracts_dir, target_dir, package_name, subdir, rest_of_path
+        )
 
-        identifier_definitions[dest_path], function_definitions[dest_path] = parse_content_for_identifier_definitions(
-            content)
+        identifier_definitions[dest_path], function_definitions[dest_path] = (
+            parse_content_for_identifier_definitions(content)
+        )
 
         imports_in_source_file[dest_path] = get_imported_identifiers_in_source_file(
-            os.path.join(contracts_dir, source_file))
+            os.path.join(contracts_dir, source_file)
+        )
 
-        if flattened_file_match:
-            additional_pragma_from_source, license[dest_path] = get_pragma_and_license_from_source_file(
-                os.path.join(contracts_dir, source_file), parse_pragma_versions=False)
-            pragma[dest_path] = block['pragma'] + '\n' + \
-                '\n'.join(additional_pragma_from_source)
-        else:
-            complete_pragma_from_source, license[dest_path] = get_pragma_and_license_from_source_file(
-                os.path.join(contracts_dir, source_file), parse_pragma_versions=True)
-            pragma[dest_path] = '\n'.join(complete_pragma_from_source)
+        pragma_version, additional_pragma, license_from_source = (
+            get_pragma_and_license_from_source_file(
+                os.path.join(contracts_dir, source_file)
+            )
+        )
 
-        # if the current file isn't the recognized source file, it will be/has been processed separately (guaranteed)
-        if interface_match and dest_path != flattened_file:
-            continue
+        license[dest_path] = license_from_source
 
+        # If source flattened file is interface or type, we want to make sure pragma version gets overwritten
+        if overwrite_pragma:
+            if normalize_pragma_version(pragma_version) != "pragma solidity >=0.5.0;":
+                pragma_version = "pragma solidity >=0.6.2;"
+
+
+        pragma[dest_path] = "\n".join([pragma_version] + additional_pragma)
         blocks[dest_path] = content
 
     for dest_path, block in blocks.items():
-        write_file(dest_path, block, license[dest_path],
-                   pragma[dest_path], identifier_definitions, function_definitions, imports_in_source_file[dest_path])
+        write_file(
+            dest_path,
+            block,
+            license[dest_path],
+            pragma[dest_path],
+            identifier_definitions,
+            function_definitions,
+            imports_in_source_file[dest_path],
+            overwrite_pragma,
+        )
 
 
-def process_all_files_in_directory(flattened_dir, contracts_dir):
+def process_all_files_in_directory(flattened_dir, contracts_dir, target_dir):
     for root, _, files in os.walk(flattened_dir):
         for file in files:
-            if file.endswith('.sol'):
+            if file.endswith(".sol"):
                 filepath = os.path.join(root, file)
-                process_file(filepath, flattened_dir,
-                             contracts_dir)
+                process_file(filepath, contracts_dir, target_dir)
 
 
 if __name__ == "__main__":
     flattened_dir = sys.argv[1]
     contracts_dir = sys.argv[2]
-    if not flattened_dir.endswith('/'):
-        flattened_dir += '/'
-    if not contracts_dir.endswith('/'):
-        contracts_dir += '/'
+    target_dir = sys.argv[3]
+    if not flattened_dir.endswith("/"):
+        flattened_dir += "/"
+    if not contracts_dir.endswith("/"):
+        contracts_dir += "/"
+    if not target_dir.endswith("/"):
+        target_dir += "/"
 
-    process_all_files_in_directory(flattened_dir, contracts_dir)
+    process_all_files_in_directory(flattened_dir, contracts_dir, target_dir)
